+++ Source/WebCore/svg/properties/SVGAnimatedEnumerationPropertyTearOff.h
@@ -28,29 +28,29 @@
 namespace WebCore {
 
 template<typename EnumType>
-class SVGAnimatedEnumerationPropertyTearOff : public SVGAnimatedStaticPropertyTearOff<int> {
+class SVGAnimatedEnumerationPropertyTearOff : public SVGAnimatedStaticPropertyTearOff<unsigned short> {
 public:
-    virtual void setBaseVal(const int& property, ExceptionCode& ec)
+    virtual void setBaseVal(const unsigned short& property, ExceptionCode& ec)
     {
         // All SVG enumeration values, that are allowed to be set via SVG DOM start with 1, 0 corresponds to unknown and is not settable through SVG DOM.
         if (property <= 0 || property > SVGPropertyTraits<EnumType>::highestEnumValue()) {
             ec = SVGException::SVG_INVALID_VALUE_ERR;
             return;
         }
-        SVGAnimatedStaticPropertyTearOff<int>::setBaseVal(property, ec);
+        SVGAnimatedStaticPropertyTearOff<unsigned short>::setBaseVal(property, ec);
     }
 
     static PassRefPtr<SVGAnimatedEnumerationPropertyTearOff<EnumType> > create(SVGElement* contextElement, const QualifiedName& attributeName, AnimatedPropertyType animatedPropertyType, EnumType& property)
     {
         ASSERT(contextElement);
-        return adoptRef(new SVGAnimatedEnumerationPropertyTearOff<EnumType>(contextElement, attributeName, animatedPropertyType, reinterpret_cast<int&>(property)));
+        return adoptRef(new SVGAnimatedEnumerationPropertyTearOff<EnumType>(contextElement, attributeName, animatedPropertyType, reinterpret_cast<unsigned short&>(property)));
     }
 
-    EnumType& currentAnimatedValue() { return reinterpret_cast<EnumType&>(SVGAnimatedStaticPropertyTearOff<int>::currentAnimatedValue()); }
+    EnumType& currentAnimatedValue() { return reinterpret_cast<EnumType&>(SVGAnimatedStaticPropertyTearOff<unsigned short>::currentAnimatedValue()); }
 
 private:
-    SVGAnimatedEnumerationPropertyTearOff(SVGElement* contextElement, const QualifiedName& attributeName, AnimatedPropertyType animatedPropertyType, int& property)
-        : SVGAnimatedStaticPropertyTearOff<int>(contextElement, attributeName, animatedPropertyType, property)
+    SVGAnimatedEnumerationPropertyTearOff(SVGElement* contextElement, const QualifiedName& attributeName, AnimatedPropertyType animatedPropertyType, unsigned short& property)
+        : SVGAnimatedStaticPropertyTearOff<unsigned short>(contextElement, attributeName, animatedPropertyType, property)
     {
     }
 };
